requirements.txt
streamlit
langchain
chromadb
transformers
torch
sentence-transformers
python-dotenv
pypdf